refactor client context type
