+++ packages/mcp/src/lib/encryption.ts
@@ -1,5 +1,6 @@
 import { createCipheriv, randomBytes } from "crypto";
 import { SERVER_VERSION } from "./constants.js";
+import type { ClientContext } from "./types.js";
 
 const DEFAULT_ENCRYPTION_KEY = "000102030405060708090a0b0c0d0e0f101112131415161718191a1b1c1d1e1f";
 const ENCRYPTION_KEY = process.env.CLIENT_IP_ENCRYPTION_KEY || DEFAULT_ENCRYPTION_KEY;
@@ -26,19 +27,6 @@ function encryptClientIp(clientIp: string): string {
     console.error("Error encrypting client IP:", error);
     return clientIp; // Fallback to unencrypted
   }
-}
-
-export interface ClientContext {
-  clientIp?: string;
-  apiKey?: string;
-  clientInfo?: {
-    ide?: string;
-    version?: string;
-  };
-  transport?: "stdio" | "http";
-  /** Mutable: set by the upstream API layer when the backend signals the
-   *  client should be prompted to sign in. Read by the auth-prompt wrapper. */
-  shouldPrompt?: boolean;
 }
 
 /**
